chore: Extract create_project in its own module
Part of #2

End of cleanup: extracted project creation logic of new command in its
own module. It avoid having too large files

*No functional changes expected*
