+++ src/cli/new.rs
@@ -16,9 +16,9 @@
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 use crate::cli::Cli;
-use crate::errors::{GenericError, Result};
+use crate::errors::Result;
+use crate::new::create_project;
 use clap::Args;
-use std::process;
 
 #[derive(Args)]
 pub struct CommandNew {
@@ -39,40 +39,12 @@ pub fn run(_cli: &Cli, command: &CommandNew, filesystem: &vfs::path::VfsPath) ->
 
     cliclack::log::success("Let's create an awesome project 🤘")?;
 
-    let name = if let Some(given_name) = &command.name {
-        cliclack::log::step(format!(
-            "Project will be created with name: {}",
-            console::style(given_name).bold()
-        ))?;
-        given_name
-    } else {
-        &cliclack::input("How do you want to call it?")
-            .placeholder("blazing-fast-forward")
-            .validate(|input: &String| {
-                if input.is_empty() {
-                    Err("Please enter a valid name")
-                } else {
-                    Ok(())
-                }
-            })
-            .interact()?
-    };
-
-    let git = if let Some(given_git) = &command.git {
-        if *given_git {
-            cliclack::log::step(format!(
-                "{} will be called",
-                console::style("git init").bold()
-            ))?;
-        }
-        given_git
-    } else {
-        &cliclack::confirm("Do you want to init git?").interact()?
-    };
+    let name = get_name(&command)?;
+    let git = get_git(&command)?;
 
     let spinner = cliclack::spinner();
     spinner.start("Initializing the project");
-    create_project(name, git, &filesystem).and_then(|_| {
+    create_project(&name, &git, &filesystem).and_then(|_| {
         spinner.stop("Done!");
 
         cliclack::note(
@@ -99,83 +71,44 @@ pub fn run(_cli: &Cli, command: &CommandNew, filesystem: &vfs::path::VfsPath) ->
     })
 }
 
-fn create_project(name: &String, git: &bool, filesystem: &vfs::path::VfsPath) -> Result<()> {
-    let name = sanitize_name(name);
-    let path = if name.starts_with("/") {
-        filesystem.root().join(&name)?
+fn get_name(command: &&CommandNew) -> Result<String> {
+    if let Some(given_name) = &command.name {
+        cliclack::log::step(format!(
+            "Project will be created with name: {}",
+            console::style(given_name).bold()
+        ))?;
+        Ok(given_name.clone())
     } else {
-        filesystem
-            .join(std::env::current_dir()?.to_str().unwrap())?
-            .join(&name)?
-    };
-    let name = if name.contains("/") {
-        path.filename()
-    } else {
-        name.clone()
-    };
-
-    check_path(&path)
-        .and_then(|_| {
-            path.create_dir_all()
-                .map_err(|err| GenericError::new(err.to_string().as_str()).into())
-        })
-        .and_then(|_| {
-            path.join("package.toml")
-                .map_err(|err| GenericError::new(err.to_string().as_str()).into())
-                .and_then(|file| {
-                    file.create_file()
-                        .map_err(|err| GenericError::new(err.to_string().as_str()).into())
-                })
-                .and_then(|mut file_stream| {
-                    write!(file_stream, "[package]\nname = {}", name)
-                        .map_err(|err| GenericError::new(err.to_string().as_str()).into())
-                })
-        })
-        .and_then(|_| {
-            if *git {
-                process::Command::new("git")
-                    .args(vec!["init", path.as_str()])
-                    .output()
-                    .map_err(|err| GenericError::new(err.to_string().as_str()).into())
-                    .and(Ok(()))
-            } else {
-                Ok(())
-            }
-        })
+        Ok(cliclack::input("How do you want to call it?")
+            .placeholder("blazing-fast-forward")
+            .validate(|input: &String| {
+                if input.is_empty() {
+                    Err("Please enter a valid name")
+                } else {
+                    Ok(())
+                }
+            })
+            .interact()?)
+    }
 }
 
-fn check_path(path: &vfs::path::VfsPath) -> Result<()> {
-    path.exists()
-        .map_err(|err| GenericError::new(err.to_string().as_str()).into())
-        .and_then(|exists| {
-            if exists {
-                path.read_dir()
-                    .map_err(|err| GenericError::new(err.to_string().as_str()).into())
-                    .and_then(|read_dir| {
-                        if read_dir.count() > 0 {
-                            Err(GenericError::new(
-                                format!("Directory {} is not empty", path.as_str()).as_str(),
-                            )
-                            .into())
-                        } else {
-                            Ok(())
-                        }
-                    })
-            } else {
-                Ok(())
-            }
-        })
-}
-
-fn sanitize_name(name: &String) -> String {
-    let parts: Vec<_> = name.trim().split_whitespace().collect();
-    parts.join("-").replace("*", "-")
+fn get_git(command: &CommandNew) -> Result<bool> {
+    if let Some(given_git) = command.git {
+        if given_git {
+            cliclack::log::step(format!(
+                "{} will be called",
+                console::style("git init").bold()
+            ))?;
+        }
+        Ok(given_git.clone())
+    } else {
+        Ok(cliclack::confirm("Do you want to init git?").interact()?)
+    }
 }
 
 #[cfg(test)]
 mod test {
     use crate::cli::new::CommandNew;
-    use crate::cli::new::sanitize_name;
     use crate::cli::{Cli, Command, new};
     use pretty_assertions::assert_eq;
     use std::io::Read;
@@ -229,17 +162,5 @@ name = {}",
             ),
             package_content
         );
-    }
-
-    #[test]
-    fn test_sanitize_name() {
-        assert_eq!("foo", sanitize_name(&"foo".to_string()));
-        assert_eq!("foo-bar", sanitize_name(&"foo bar".to_string()));
-        assert_eq!("foo-bar", sanitize_name(&"foo-bar".to_string()));
-        assert_eq!("foo_bar", sanitize_name(&"foo_bar".to_string()));
-        assert_eq!("foo", sanitize_name(&"   foo   ".to_string()));
-        assert_eq!("foo-bar", sanitize_name(&"   foo   bar   ".to_string()));
-        assert_eq!("foo&bar", sanitize_name(&"foo&bar".to_string()));
-        assert_eq!("foo-bar", sanitize_name(&"foo*bar".to_string()));
     }
 }
